refactor(cdrinput, cdrlabelstandalone): fix v-model and styling issues with CdrInput. Add uids
Refactor CdrInput to use <script setup> syntax. Fixed v-model issue by referencing a new computed
‘inputModel’ with its own custom getter and setter. Ensured all fallthrough attributes except class
get applied to the input element. The class attribute is bound to the outermost element. Added a uid
module to provide a unique id for each component should an id not be provided by the user. Fixed
some implementations issue on the CdrInput example page. Fixed a typo in CdrLabelStandalone.