refactor(tabs): rewrite of tabs for vue 3
almost a complete rewrite of the tabs component. replaced <nav> and <div> markup with <ul> and <li>
markup. most styling remains unchanged, with a few changes to account for the different markup. most
of the same css classes are used. logic dealing with animations and overflow gradient have been
copy/pasted from the old version of tabs and integrated into this new verison. some refactoring of
the old tests have been done, but not a complete overhaul--just ensuring that all the existing tests
still test what they had in the past.