refactor(cdr-tabs and cdr-tab-panel): rewrite cdr-tabs for vue 3
Almost a complete rewrite of the tabs component. Changed markup to be more semantically appropriate.
Most styling remains unchanged, with a few changes to account for the different markup. Most of the
same CSS classes are used. Logic dealing with animations and overflowgradient have been copy/pasted
from the old version of tabs and integrated into this new version. Some refactoring of the old tests
have been done, but it's far from a complete overhaul--just ensuring that all the existing tests
still test what they had done in the past.